Modified container header to auto-adjust based on header view
