+++ Kwartracker/App Main/KwartrackerApp.swift
@@ -15,7 +15,7 @@ struct KwartrackerApp: App {
 
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            SignInView()
         }
     }
 }
