+++ Kwartracker/Scenes/Sign-in Page/Content/HeaderView.swift
@@ -1,0 +1,28 @@
+//
+//  SwiftUIView.swift
+//  Kwartracker
+//
+//  Created by John Ellie Go on 4/14/21.
+//
+
+import SwiftUI
+
+struct HeaderView: View {
+    var body: some View {
+        VStack {
+            HStack(spacing: 10) {
+                Image(Asset.Images.headerLogoIcon.name)
+                Spacer()
+            }
+            .padding(.leading, 50)
+        }
+    }
+}
+
+struct HeaderView_Previews: PreviewProvider {
+    static var previews: some View {
+        HeaderView()
+            .background(Color("Teal"))
+            .previewLayout(.sizeThatFits)
+    }
+}
